backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pydantic
PyPDF2
pdfplumber
rapidfuzz
google-genai
python-dotenv
